//bringing things in line to work with 2.5 (mostly admin changes)
git-svn-id: http://plugins.svn.wordpress.org/organize-series/trunk@34615 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ orgSeries/series-taxonomy.php
@@ -21,6 +21,10 @@ function _usort_series_by_part($a, $b) {
 		return 0;
 	}
 
+function _usort_series_by_name($a, $b) {
+	return strcmp( $a['ser_name'], $b['ser_name'] );
+}
+
 	
 //This function is used to create an array of posts in a series including the order the posts are in the series.  Then it will sort the array so it is keyed in the order the posts are in.  Will return the array.
 function get_series_order ($posts, $postid = 0, $skip = TRUE) {
@@ -102,9 +106,10 @@ function series_createRewriteRules($rules) {
 	$series_structure = $wp_rewrite->root . SERIES_QUERYVAR . "/$series_token";
 	$rewrite = $wp_rewrite->generate_rewrite_rules($series_structure);
 	//return $series_structure;
+	
 	return ( $rewrite + $rules );
-
 }
+
 function series_init() {
 	global $wp_rewrite;
 	
@@ -126,14 +131,9 @@ function series_init() {
 	if ($series_toc_url && (strpos($_SERVER['REQUEST_URI'], $series_toc_url) === 0) && (strlen($_SERVER['REQUEST_URI']) == strlen($series_toc_url))) {
 		status_header(200); 
 		add_filter('request', 'orgSeries_request');
-		add_action('parse_query', 'orgSeries_parse_query');
 		add_action('template_redirect', 'orgSeries_toc_template');
 	}
 	//$wp_rewrite->flush_rules();  //TODO: (NEEDS TESTING) I THINK THIS MIGHT MAKE IT SO USERS DON'T HAVE TO UPDATE THE PERMALINKS MANUALLY.  Probably should be added to the plugin_activation hook.
-}
-
-function orgSeries_parse_query($wp_query) {
-	$wp_query->is_404 = false;
 }
 
 function orgSeries_request($query_vars) {
@@ -174,14 +174,14 @@ function get_series_link( $series_id ) {
 		$serieslink = $file . '?series=' . $id;
 	} else {
 		$serieslink = str_replace($series_token, $slug, $serieslink);
-		$serieslink = get_settings('home') . user_trailingslashit($serieslink, 'series');
+		$serieslink = get_option('home') . user_trailingslashit($serieslink, 'series');
 	}
 	return $serieslink;
 	//return apply_filters('series_link', $serieslink, $series_id); 
 }
 	
 function get_the_series( $id = false ) { 
-	global $post, $tem_cache, $blog_id;
+	global $post, $term_cache, $blog_id;
 	
 	$id = (int) $id;
 	
@@ -590,6 +590,7 @@ function wp_dropdown_series($args = '') {
 
 function walk_series_dropdown_tree($serieslist, $r) {
 	$series_name = apply_filters('list_series', $serieslist->name, $serieslist);
+	$output = '';
 	$output .= "\t<option value=\"" . $serieslist->term_id . "\"";
 	if ( $serieslist->term_id == $r['selected'] )
 		$output .= ' selected="selected"';
@@ -736,13 +737,14 @@ function series_addQueryVar($wpvar_array) {
 	$wpvar_array[] = SERIES_QUERYVAR;
 	return($wpvar_array);
 }
+
 //for series queries
 add_filter('query_vars', 'series_addQueryVar');
 add_action('parse_query','series_parseQuery');
 
 
-function series_parseQuery($query) {
-	//if this is a series query, then reset other is_x flags and add query filters
+function series_parseQuery() {
+	//if this is a series query, then reset other is_x flags and add template redirect;
 	if (is_series()) {
 		global $wp_query;
 		$wp_query->is_single = false;
@@ -751,11 +753,12 @@ function series_parseQuery($query) {
 		$wp_query->is_search = false;
 		$wp_query->is_home = false;
 		$wp_query->is_series = true;
+		$wp_query->is_404 = false;
 		
-		add_filter('posts_where', 'series_postsWhere');
-		add_filter('posts_join', 'series_postsJoin');
 		add_action('template_redirect','series_includeTemplate');
 	}	
+	add_filter('posts_where', 'series_postsWhere');
+	add_filter('posts_join', 'series_postsJoin');
 }
 
 function is_series() { 
@@ -771,6 +774,7 @@ function is_series() {
 function series_postsWhere($where) { 
 	global $wpdb;
 	$series_var = get_query_var(SERIES_QUERYVAR);
+	$cat_var = get_query_var('cat');
 	$token = "'" . SERIES_QUERYVAR . "'";
 	//convert to series id if permalinks turned on.
 	$serchk = is_term( $series_var, SERIES_QUERYVAR );
@@ -778,24 +782,40 @@ function series_postsWhere($where) {
 		$series_var = $serchk['term_id'];
 	$whichseries = '';
 	
-	if ( !empty($series_var) ) {
+	if ( !empty($series_var)  && empty($cat_var) ) {
 		$whichseries .= " AND $wpdb->term_taxonomy.taxonomy = $token ";
 		$whichseries .= " AND $wpdb->term_taxonomy.term_id = $series_var ";
 	}
 		
-	$where .= "AND $wpdb->term_taxonomy.taxonomy = $token ";
+	//for category and series intersects
+	If ( !empty( $series_var ) && !empty($cat_var) ) {
+		$taxonomy = $token;
+		$t_ids = array( $cat_var, $series_var );
+		$tsql = "SELECT p.ID FROM $wpdb->posts p INNER JOIN $wpdb->term_relationships tr ON (p.ID = tr.object_id) INNER JOIN $wpdb->term_taxonomy tt ON (tr.term_taxonomy_id = tt.term_taxonomy_id) INNER JOIN $wpdb->terms t ON (tt.term_id = t.term_id)";
+		$tsql .= " WHERE tt.taxonomy = ($token OR 'category') AND t.term_id IN ('" . implode("', '", $t_ids) . "')";
+		$tsql .= " GROUP BY p.ID HAVING count(p.ID) = " . count($t_ids);
+		
+		$post_ids = $wpdb->get_col($tsql);
+		
+		if ( count($post_ids) )
+			$whichseries .= " AND $wpdb->posts.ID IN (" . implode(', ', $post_ids) . ") ";
+		else 
+			$whichseries = " AND 0 = 1";
+	}
+			
 	$where .= $whichseries;
-	return ($where);
+	return $where;
 }
 
 function series_postsJoin($join) {
 	global $wpdb;
 	$series_var = get_query_var(SERIES_QUERYVAR);
-	if ( !empty($series_var) )  {
-		$join = " LEFT JOIN $wpdb->term_relationships ON ($wpdb->posts.ID = $wpdb->term_relationships.object_id) LEFT JOIN $wpdb->term_taxonomy ON ($wpdb->term_relationships.term_taxonomy_id = $wpdb->term_taxonomy.term_taxonomy_id) ";
+	$cat_var = get_query_var('cat');
+	if ( !empty($series_var) && empty( $cat_var ) )  {
+		$join = " INNER JOIN $wpdb->term_relationships ON ($wpdb->posts.ID = $wpdb->term_relationships.object_id) INNER JOIN $wpdb->term_taxonomy ON ($wpdb->term_relationships.term_taxonomy_id = $wpdb->term_taxonomy.term_taxonomy_id) ";
 	}
 		
-	return ($join);	
+	return $join;	
 }
 
 function series_includeTemplate() {
@@ -846,12 +866,14 @@ function wp_set_post_series( $post_ID = 0, $series_id = 0) {
 function wp_delete_post_series_relationship( $id = 0 ) {
 	global $wpdb, $wp_rewrite;
 	$postid = (int) $id;
-	$series = get_the_series($postid);
+	if ( $series = get_the_series($postid) ) {
 	$seriesid = $series->term_id;
 	delete_post_meta($postid, SERIES_PART_KEY);
 	$success = wp_delete_object_term_relationships($postid, array('series'));
 	if ( $success ) return wp_reset_series_order_meta_cache($postid, $seriesid);
-	else return FALSE;
+	else return false;
+	}
+	return FALSE;
 }
 
 //add_action('edit_post','wp_set_post_series');
@@ -929,15 +951,18 @@ function wp_insert_series($serarr, $file = FALSE) {
 	$overrides = array('action' => $action);
 	if (!($file) || $file=='') unset($file);
 	
-	if (isset($file))
+	if (isset($file)) {
 		$iconfile = wp_handle_upload( $file, $overrides );
 	
-	//if ($message = $iconfile['error']) return FALSE; //TODO - remove the RETURN FALSE check and instead return an array for wp_insert_series containing $message, and $series_id.  This would require going back over all the files to update any calls to wp_insert_series so that returned variable is used correctly.
-	$iconname = $iconfile['url'];
-	
-	//take the $iconname which contains the full url of the series
-	$iconname = explode('/', $iconname);
-	$icon = $iconname[count($iconname) - 1];
+		//if ($message = $iconfile['error']) return FALSE; //TODO - remove the RETURN FALSE check and instead return an array for wp_insert_series containing $message, and $series_id.  This would require going back over all the files to update any calls to wp_insert_series so that returned variable is used correctly.
+		$iconname = $iconfile['url'];
+		
+		//take the $iconname which contains the full url of the series
+		$iconname = explode('/', $iconname);
+		$icon = $iconname[count($iconname) - 1];
+	} else {
+		$icon = '';
+	}
 	
 	$args = compact('name','slug','description');
 	
@@ -991,7 +1016,7 @@ function series_rows( $series = 0 ) {
 }
 
 function _series_row($series) {
-	global $class;
+	global $class, $wp_version;
 	
 	$series_icon = series_get_icons($series->term_id);
 	$series_url = seriesicons_url();
@@ -999,11 +1024,18 @@ function _series_row($series) {
 	
 	if ( current_user_can( 'manage_series' ) ) {
 		$edit = "<a href='edit.php?page=orgSeries/orgSeries-manage.php&amp;action=edit&amp;series_ID=$series->term_id' class='edit'>".__( 'Edit' )."</a></td>";
-		$edit .= "<td><a href='" . wp_nonce_url("edit.php?page=orgSeries/orgSeries-manage.php&amp;action=delete&amp;series_ID=$series->term_id", 'delete-series_' . $series->term_id ) . "' onclick=\"return deleteSomething('serial', $series->term_id, '" . js_escape(sprintf( __("You are about to delete the series '%s'. \nAll posts that were assigned to this series will be disassociated from the series.\n'OK' to delete, 'Cancel' to stop." ), $series->name  )) . "' );\" class='delete'>".__( 'Delete' )."</a>";
+		
+		if ( isset( $wp_version ) && $wp_version >= 2.5 )
+			$edit .=  "<td><a href='" . wp_nonce_url("edit.php?page=orgSeries/orgSeries-manage.php&action=delete&amp;series_ID=$series->term_id&noheader=1", 'delete-series_' . $series->term_id ) . "' class='delete'>".__( 'Delete' )."</a>";
+		else
+			$edit .= "<td><a href='" . wp_nonce_url("edit.php?page=orgSeries/orgSeries-manage.php&amp;action=delete&amp;series_ID=$series->term_id", 'delete-series_' . $series->term_id ) . "' onclick=\"return deleteSomething('serial', $series->term_id, '" . js_escape(sprintf( __("You are about to delete the series '%s'. \nAll posts that were assigned to this series will be disassociated from the series.\n'OK' to delete, 'Cancel' to stop." ), $series->name  )) . "' );\" class='delete'>".__( 'Delete' )."</a>";
 	} else
 		$edit = '';
 	
-	$class = ( ( defined( 'DOING_AJAX' ) && DOING_AJAX ) || " class='alternate'" == $class ) ? '' : " class='alternate'";
+	if ( isset( $wp_version ) && $wp_version >= 2.5 )
+		$class = " class='alternate'" == $class ? '' : " class='alternate'";
+	else
+		$class = ( ( defined( 'DOING_AJAX' ) && DOING_AJAX ) || " class='alternate'" == $class ) ? '' : " class='alternate'";
 	$series->count = number_format_i18n( $series->count );
 	$posts_count = ( $series->count > 0 ) ? "<a href='edit.php?series=$series->term_id'>$series->count</a>" : $series->count;  
 	$output = "<tr id='serial-$series->term_id'$class>
